requirements.txt
discord.py
requests
aiohttp
python-dotenv
SQLAlchemy>=2.0
